+++ src/main.ts
@@ -7,15 +7,15 @@ import TheUpdater from "./functions/updater.ts";
 import TheSettings from "./commands/settings.ts";
 
 import { I_LIKE_JS, VERSION } from "./constants.ts";
-import type { CONFIG_FILES, MANAGERS } from "./types.ts";
+import type { CONFIG_FILES } from "./types.ts";
 import { LogStuff, ParseFlag } from "./functions/io.ts";
 import { FreshSetup } from "./functions/config.ts";
 
 const [inputCommand] = Deno.args;
 
 if (!inputCommand) {
-    await init();
-    await TheHelper();
+    await init(false);
+    await TheHelper({});
     Deno.exit(0);
 }
 
@@ -26,18 +26,18 @@ const flags = Deno.args.map((arg) => {
 const isVerbose = flags.includes("--verbose");
 const wantsToUpdate = flags.includes("--update");
 
-async function init(update?: boolean) {
+async function init(update: boolean) {
     const configFiles = await FreshSetup(); // Temporarily hold the result
     const ALL_CONFIG_FILES: CONFIG_FILES = configFiles; // Assign only after it's resolved
     await TheUpdater(ALL_CONFIG_FILES, update);
     return ALL_CONFIG_FILES;
 }
 
-const ALL_CONFIG_FILES = await init();
+const ALL_CONFIG_FILES = await init(false);
 
 if (ParseFlag("help", true).some((flag) => flags.includes(flag))) {
-    await init();
-    await TheHelper(Deno.args[1]);
+    await init(false);
+    await TheHelper({ query: Deno.args[1] });
     Deno.exit(0);
 }
 
@@ -46,52 +46,56 @@ if (ParseFlag("version", true).some((flag) => flags.includes(flag))) {
     Deno.exit(0);
 }
 
-if (ParseFlag("experimental-stats", false).some((flag) => flags.includes(flag))) {
-    await init();
-    if (confirm("Experimental features are hidden behind a flag for a reason. Are you sure?")) {
-        await LogStuff(`Okay, there we go. Report any ${I_LIKE_JS.FKN} error you find in GitHub.`);
+if (
+    ParseFlag("experimental-stats", false).some((flag) => flags.includes(flag))
+) {
+    await init(false);
+    if (
+        confirm(
+            "Experimental features are hidden behind a flag for a reason. Are you sure?",
+        )
+    ) {
+        await LogStuff(
+            `Okay, there we go. Report any ${I_LIKE_JS.FKN} error you find in GitHub.`,
+        );
         await TheStatistics(ALL_CONFIG_FILES);
     }
     Deno.exit(0);
 }
 
 async function main(command: string) {
+    await init(false);
+
     try {
         switch (command.toLowerCase()) {
             case "clean":
-                await init();
-                if (Deno.args[1] && !["normal", "hard", "maxim"].includes(Deno.args[1])) {
-                    throw new Error("Invalid intensity provided.");
-                }
-                await TheCleaner(isVerbose, wantsToUpdate, Deno.args[1] as "normal" | "hard" | "maxim" ?? "normal", ALL_CONFIG_FILES);
+                await TheCleaner({
+                    verbose: isVerbose,
+                    update: wantsToUpdate,
+                    intensity: Deno.args[1] ?? "normal",
+                    CF: ALL_CONFIG_FILES,
+                });
                 break;
             case "manager":
-                await init();
                 await TheManager(Deno.args, ALL_CONFIG_FILES);
                 break;
             case "settings":
-                await init();
-                await TheSettings(Deno.args, ALL_CONFIG_FILES);
+                await TheSettings({ args: Deno.args, CF: ALL_CONFIG_FILES });
                 break;
             case "migrate":
-                await init();
-                if (!Deno.args[1]) throw new Error(`No project specified!`);
-                if (!Deno.args[2]) throw new Error(`No target specified!`);
-                await TheMigrator(Deno.args[1], Deno.args[2] as MANAGERS);
+                await TheMigrator({ project: Deno.args[1], target: Deno.args[2] });
                 break;
             case "self-update":
                 await init(true);
                 break;
             case "stats":
-                await init();
                 await LogStuff(
                     "stats has been proven to be an unstable and not-working command. Use --experimental-stats if you want to try it.",
                     "warn",
                 );
                 break;
             default:
-                await init();
-                await TheHelper();
+                await TheHelper({});
                 Deno.exit(0);
         }
     } catch (e) {