+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - Added partial support for cleanup of both the Bun and Deno JavaScript runtimes.
 - Added the ability to customize whether an ignored project should ignore updates, cleanup, or everything.
   - Now that the ignore file expects content, comments are allowed using the `--!` prefix.
+- Added better support for workspaces, by recognizing `pnpm-workspace.yaml`.
 
 ### Changed